# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TuMobilya-API",
        "description": "TuMobilya-API için kimlik doğrulama ve mobilya yönetimi API'si",
        "contact": {
            "email": "info@tumobilya.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Yerel geliştirme sunucusu"
        },
        {
            "url": "http://tumobilya-api.test"
        }
    ],
    "paths": {
...
        "/api/sliders": {
            "get": {
                "tags": [
                    "Slaytlar"
                ],
                "summary": "Tüm slaytları listele",
                "description": "Kayıtlı tüm slaytları döndürür.",
                "operationId": "f3f821bb23189c1aa1f07dcc34fbd7b3",
                "responses": {
                    "200": {
                        "description": "Slaytlar başarıyla listelendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Yeni Koleksiyon"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "2025 mobilya koleksiyonumuzu keşfedin!"
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "slider1.jpg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Kimlik doğrulanmadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Slaytlar"
                ],
                "summary": "Yeni slayt ekle",
                "description": "Yeni bir slayt oluşturur.",
                "operationId": "40e04352099ba3630b0ecd3d8a07ff4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Yaz İndirimi"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Tüm ürünlerde %20 indirim!"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "slider3.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Slayt başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Yaz İndirimi"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Tüm ürünlerde %20 indirim!"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "slider3.jpg"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Slayt oluşturuldu"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasyon hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The title field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sliders/{id}": {
            "get": {
                "tags": [
                    "Slaytlar"
                ],
                "summary": "Tek bir slaytı getir",
                "description": "Belirtilen ID'ye sahip slaytı döndürür.",
                "operationId": "3bfeb31fa1033fda85c708993d64ea25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slayt bulundu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Yeni Koleksiyon"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "2025 mobilya koleksiyonumuzu keşfedin!"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "slider1.jpg"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slayt bulunamadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kaynak bulunamadı"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Slaytlar"
                ],
                "summary": "Slayt güncelle",
                "description": "Belirtilen ID'ye sahip slaytı günceller.",
                "operationId": "69287588ebbfccfab3ba79f95438e0df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Yaz İndirimi (Güncellendi)"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Tüm ürünlerde %25 indirim!"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "slider3-updated.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Slayt güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Yaz İndirimi (Güncellendi)"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Tüm ürünlerde %25 indirim!"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "slider3-updated.jpg"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Slayt güncellendi"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slayt bulunamadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kaynak bulunamadı"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Slaytlar"
                ],
                "summary": "Slayt sil",
                "description": "Belirtilen ID'ye sahip slaytı siler.",
                "operationId": "47b0483bc8d101c4a10bea5ef85ea075",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slayt silindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Slayt silindi"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slayt bulunamadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kaynak bulunamadı"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}